moved static function to helper from basid data servlet
